+++ app/src/components/Icon.vue
@@ -13,6 +13,8 @@ const paths = {
   up: "M12 19V5M5 12l7-7 7 7",
   folder: "M3 7a2 2 0 0 1 2-2h4l2 2h8a2 2 0 0 1 2 2v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V7z",
   file: "M14 3H7a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V8l-5-5zM14 3v5h5",
+  stop: "M6 6h12v12H6z",
+  chevron: "M9 6l6 6-6 6",
 };
 </script>